+++ modules/seqexec/engine/src/main/scala/seqexec/engine/package.scala
@@ -7,7 +7,7 @@ import cats.data.Kleisli
 import cats.effect.IO
 import seqexec.engine.Result.{Error, PartialVal, PauseContext, RetVal}
 import seqexec.model.ActionType
-import seqexec.model.Model.Conditions
+import seqexec.model.Conditions
 import seqexec.model.{ Observer, Operator}
 
 package engine {
