+++ modules/seqexec/server/src/test/scala/seqexec/server/SeqexecEngineSpec.scala
@@ -21,9 +21,8 @@ import seqexec.engine._
 import seqexec.server.SeqexecEngine.Settings
 import seqexec.server.keywords.GDSClient
 import seqexec.model.enum.{ ActionStatus, CloudCover, ImageQuality, Instrument, Resource, SkyBackground, WaterVapor}
-import seqexec.model.Model.{ Conditions, SequenceMetadata }
-import seqexec.model.{ Operator, SequenceState }
-import seqexec.model.{ActionType, UserDetails}
+import seqexec.model.Model.{ SequenceMetadata }
+import seqexec.model._
 
 @SuppressWarnings(Array("org.wartremover.warts.NonUnitStatements"))
 @SuppressWarnings(Array("org.wartremover.warts.Throw"))
