+++ modules/seqexec/server/src/test/scala/seqexec/server/SeqexecServerArbitraries.scala
@@ -21,7 +21,7 @@ import seqexec.server.gcal.GcalController._
 import seqexec.server.tcs.{CRFollow, TcsController, TcsControllerEpics}
 import seqexec.server.keywords._
 import seqexec.model.enum.Instrument
-import seqexec.model.Model.Conditions
+import seqexec.model.Conditions
 import seqexec.model.Operator
 import edu.gemini.spModel.gemini.flamingos2.Flamingos2
 import edu.gemini.spModel.gemini.gnirs.GNIRSParams
