+++ modules/seqexec/web/server/src/main/scala/seqexec/web/server/http4s/SeqexecCommandRoutes.scala
@@ -9,8 +9,8 @@ import seqexec.server.Commands
 import seqexec.server.SeqexecEngine
 import seqexec.server
 import seqexec.model.enum.{ CloudCover, ImageQuality, SkyBackground, WaterVapor}
-import seqexec.model.Model.{ SequencesQueue, Conditions }
-import seqexec.model.{ Observer, Operator }
+import seqexec.model.Model.SequencesQueue
+import seqexec.model.{ Conditions, Observer, Operator }
 import seqexec.model.UserDetails
 import seqexec.web.server.model.CommandsModel._
 import seqexec.web.server.http4s.encoder._
