+++ modules/seqexec/web/server/src/main/scala/seqexec/web/server/http4s/encoder/BooEncoders.scala
@@ -11,7 +11,8 @@ import org.http4s.booPickle._
 import seqexec.model._
 import seqexec.model.boopickle.GemModelBooPicklers
 import seqexec.model.enum.{ CloudCover, ImageQuality, SkyBackground, WaterVapor}
-import seqexec.model.Model.{ Conditions, SequencesQueue }
+import seqexec.model.Model.SequencesQueue
+import seqexec.model.Conditions
 import seqexec.model.Operator
 import seqexec.web.common.{CliCommand, LogMessage}
 import seqexec.web.common.LogMessage._
